A lot of improvements

+++ app/Http/Controllers/Admin/ItemController.php
@@ -8,6 +8,8 @@ use App\Http\Requests\UpdateCategoryItemRequest;
 use App\Models\Category;
 use App\Models\Item;
 use App\Models\Photo;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class ItemController extends Controller
@@ -33,6 +35,10 @@ class ItemController extends Controller
 
     public function store(StoreCategoryItemRequest $request, Category $category)
     {
+        if (Gate::denies('be-admin')) {
+            throw new AuthorizationException('You are not authorized to perform this action.');
+        }
+
         $item = $category->items()->create($request->validated()); //переписати на окремі інпути
         $item->modifications()->create(['label' => $request->input('modification')]);
         if ($request->hasFile('photo')) {
@@ -46,6 +52,10 @@ class ItemController extends Controller
 
     public function edit(Category $category, Item $item)
     {
+        if (Gate::denies('be-admin')) {
+            throw new AuthorizationException('You are not authorized to perform this action.');
+        }
+        
         $categories = Category::all();
         
         return view('items.admin.edit', compact('item', 'categories', 'category'));
@@ -53,6 +63,10 @@ class ItemController extends Controller
 
     public function update(UpdateCategoryItemRequest $request, Category $category, Item $item)
     {
+        if (Gate::denies('be-admin')) {
+            throw new AuthorizationException('You are not authorized to perform this action.');
+        }
+        
         $item->update($request->validated());
 
         return redirect()->route('admin.categories.show', $category);
@@ -60,6 +74,10 @@ class ItemController extends Controller
 
     public function destroy(Category $category, Item $item)
     {
+        if (Gate::denies('be-admin')) {
+            throw new AuthorizationException('You are not authorized to perform this action.');
+        }
+
         $item->photos->each(fn($photo) => $photo->deleteFile());
         
         $item->photos()->delete();
